+++ Kernel/Core/Korlib/System/SByte.cs
@@ -13,7 +13,7 @@ using SharpOS.Kernel.ADC;
 
 namespace InternalSystem {
 	[TargetNamespace ("System")]
-	public struct Int32 {
-		internal int Value;
+	public struct SByte {
+		internal sbyte Value;
 	}
 }
