+++ Kernel/Core/Korlib/System/UInt16.cs
@@ -13,7 +13,7 @@ using SharpOS.Kernel.ADC;
 
 namespace InternalSystem {
 	[TargetNamespace ("System")]
-	public struct Int32 {
-		internal int Value;
+	public struct UInt16 {
+		internal ushort Value;
 	}
 }
